requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple

cachetools==4.2.0; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
configparser==5.0.1
google-auth==1.23.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
grpc-interceptor==0.12.0
grpcio-health-checking==1.34.0
grpcio-tools==1.34.0
grpcio==1.34.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jmespath==0.10.0
kubernetes==12.0.1
marshmallow-enum==1.5.1
marshmallow==3.9.1
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
protobuf==3.13.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.25.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.6; python_version >= '3.5'
six==1.15.0
urllib3==1.25.11
websocket-client==0.57.0

